Ejercicio de variables

+++ Tema01/Variables/Ejercicio4.java
@@ -2,7 +2,7 @@
 
 public class Ejercicio4 {
   public static void main(String[] args) {
-    double euro;
+    int euro;
     double peseta;
 
     euro = 1;
